[REF] l10n_br_account: mv tests to move_workflow
