+++ l10n_br_account/tests/__init__.py
@@ -1,5 +1,3 @@
-# License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
-
 from . import test_account_move_sn
 from . import test_account_move_lc
 from . import test_account_taxes
@@ -8,3 +6,4 @@ from . import test_move_document_discount
 from . import test_invoice_refund
 from . import test_multi_localizations_invoice
 from . import test_payment_status
+from . import test_move_workflow
