+++ l10n_br_account/tests/test_account_move_lc.py
@@ -2131,31 +2131,3 @@ class AccountMoveLucroPresumido(AccountMoveBRCommon):
         )
         self.assertEqual(len(self.move_in_compra_para_revenda.line_ids), 11)
         self.assertEqual(self.move_in_compra_para_revenda.amount_total, 2065.0)
-
-    def test_change_states(self):
-        # first we make a few assertions about an existing vendor bill:
-        document_id = self.move_out_venda.fiscal_document_id
-        self.assertEqual(self.move_out_venda.state, "draft")
-        self.assertEqual(document_id.state, "em_digitacao")
-        self.move_out_venda.action_post()
-        self.assertEqual(self.move_out_venda.state, "posted")
-        fiscal_edi = self.env["ir.module.module"].search(
-            [("name", "=", "l10n_br_fiscal_edi")]
-        )
-        if fiscal_edi and fiscal_edi.state == "installed":
-            self.assertEqual(document_id.state, "a_enviar")
-            self.move_out_venda.button_draft()
-            self.assertEqual(self.move_out_venda.state, "draft")
-            self.assertEqual(document_id.state, "em_digitacao")
-            document_id.action_document_confirm()
-            self.assertEqual(self.move_out_venda.state, "posted")
-            self.assertEqual(document_id.state, "a_enviar")
-            document_id.action_document_back2draft()
-            self.assertEqual(self.move_out_venda.state, "draft")
-            self.assertEqual(document_id.state, "em_digitacao")
-
-    def test_document_deny(self):
-        document_id = self.move_out_venda.fiscal_document_id
-        self.assertEqual(self.move_out_venda.state, "draft")
-        document_id.exec_after_SITUACAO_EDOC_DENEGADA("em_digitacao", "denegada")
-        self.assertEqual(self.move_out_venda.state, "cancel")
